requirements.txt
Flask==3.0.2
pymongo==4.6.1
joblib==1.3.2
numpy>=2.2.0
pandas>=2.2.0
requests==2.31.0
python-dotenv==1.0.1
flask-caching==2.1.0
apscheduler==3.10.4
scikit-learn>=1.6.0
flask_login>=0.6.3
flask_mail==0.9.1
PyJWT==2.8.0
Werkzeug==3.0.1